Move responses stuff to separate file.
